# control-plane/docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Control-plane is a central microservice of Service Mesh which is responsible for all the gateways configuration. For more information, visit our Documentation(https://github.com/Netcracker/qubership-core-control-plane/tree/main/README.md).",
        "title": "Control Plane API",
        "version": "2.0"
    },
    "paths": {
...
        "/api/v3/promote/{version}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Post Promote Version",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "control-plane-v3"
                ],
                "summary": "Post Promote Version",
                "operationId": "PostPromoteVersionV3",
                "parameters": [
                    {
                        "type": "string",
                        "description": "version",
                        "name": "version",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "archiveSize",
                        "name": "archiveSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.DeploymentVersion"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v3/rollback": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Post Rollback Version",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "control-plane-v3"
                ],
                "summary": "Post Rollback Version",
                "operationId": "PostRollbackVersion",
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.DeploymentVersion"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v3/versions": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get Deployment Versions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "control-plane-v3"
                ],
                "summary": "Get Deployment Versions",
                "operationId": "GetDeploymentVersionsV3",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.DeploymentVersion"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v3/versions/microservices/{microservice}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get Microservice Version",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "control-plane-v3"
                ],
                "summary": "Get Microservice Version",
                "operationId": "GetMicroserviceVersion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "microservice",
                        "name": "microservice",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v3/versions/{version}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete Deployment Version With ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "control-plane-v3"
                ],
                "summary": "Delete Deployment Version With ID",
                "operationId": "DeleteDeploymentVersionWithIDV3",
                "parameters": [
                    {
                        "type": "string",
                        "description": "version",
                        "name": "version",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.DeploymentVersion"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.DeploymentVersion": {
            "type": "object",
            "properties": {
                "createdWhen": {
                    "type": "string"
                },
                "stage": {
                    "type": "string"
                },
                "updatedWhen": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}